T157 Reorganize folders in the client
Reorganized folders in the client so that components that are currently
only used by their parent, or components under their parent, are in the
same folder. This seems to help a lot with logically understanding where
components are.
