+++ client/src/main/components/ProjectTable/CompletableRow/CompletedCheckbox.tsx
@@ -1,7 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import { Checkbox } from '@material-ui/core';
-import { CompletableType, Completable } from '../../utils/dbTypes';
-import UserData from '../../clientData/UserData';
+import { CompletableType, Completable } from '../../../utils/dbTypes';
+import UserData from '../../../clientData/UserData';
 
 export type CompletedCheckboxProps = {
   className?: string;
