+++ client/src/main/components/ProjectTable/CompletableRow/DateInput.tsx
@@ -1,8 +1,8 @@
 import React, { useState, useEffect } from 'react';
 import { MaterialUiPickersDate } from '@material-ui/pickers/typings/date';
 import { DatePicker } from '@material-ui/pickers';
-import UserData from '../../clientData/UserData';
-import { CompletableType } from '../../utils/dbTypes';
+import UserData from '../../../clientData/UserData';
+import { CompletableType } from '../../../utils/dbTypes';
 
 export type DateInputProps = {
   completableType: CompletableType;
