+++ client/src/main/components/ProjectTable/CompletableRow/NoteInput.tsx
@@ -1,8 +1,8 @@
 import React, { useState } from 'react';
 import { TextField } from '@material-ui/core';
-import { resetTimer } from '../../utils/savingTimer';
-import UserData from '../../clientData/UserData';
-import { CompletableType } from '../../utils/dbTypes';
+import { resetTimer } from '../../../utils/savingTimer';
+import UserData from '../../../clientData/UserData';
+import { CompletableType } from '../../../utils/dbTypes';
 
 export type NoteInputProps = {
   completableType: CompletableType;
