+++ client/src/main/components/ProjectTable/CompletableRow/PriorityButton/PriorityDialog.tsx
@@ -4,8 +4,8 @@ import DialogContent from '@material-ui/core/DialogContent';
 import DialogTitle from '@material-ui/core/DialogTitle';
 import { DialogActions, Button } from '@material-ui/core';
 import PriorityInput from './PriorityInput';
-import { CompletableType } from '../../utils/dbTypes';
-import UserData from '../../clientData/UserData';
+import { CompletableType } from '../../../../utils/dbTypes';
+import UserData from '../../../../clientData/UserData';
 
 type PriorityDialogProps = {
   open: boolean;
