+++ client/src/main/components/TopMenuBar/SettingsDialog/FontSizeSetting.tsx
