+++ client/src/main/components/TopMenuBar/index.tsx
@@ -22,11 +22,11 @@ import SettingsIcon from '@material-ui/icons/Settings';
 import HelpIcon from '@material-ui/icons/Help';
 import SettingsDialog from './SettingsDialog';
 import LoginDialog from './LoginDialog';
-import { AlertFunction } from '../App';
-import { logout } from '../utils/fetchMethods';
-import UserData from '../clientData/UserData';
-import { manualSave, windowUnloadListener } from '../utils/savingTimer';
-import { AppSaveStatus } from '../clientData/AppSaveStatus';
+import { AlertFunction } from '../../App';
+import { logout } from '../../utils/fetchMethods';
+import UserData from '../../clientData/UserData';
+import { manualSave, windowUnloadListener } from '../../utils/savingTimer';
+import { AppSaveStatus } from '../../clientData/AppSaveStatus';
 
 /* This eslint comment is not a good solution, but the alternative seems to be 
 ejecting from create-react-app */
